requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
bcrypt
python-dotenv